Apply change of web-nginx-mysql container's listen port
Previous zabbix-web-nginx-mysql container's listen port had been
listened tcp:80 for HTTP request. But it was changed to 8080 on 29 Apr
by change of zabbix/zabbix-docker. By this change, stackstorm-zabbix's
integration test couldn't access to a test zabbix server. This commit
applies this listen port change to be able to test properly.